API Reference

Public

Internal

GFF3.Reader — Method

GFF3.Reader(input::IO;
            index=nothing,
            save_directives::Bool=false,
            skip_features::Bool=false,
            skip_directives::Bool=true,
            skip_comments::Bool=true)

GFF3.Reader(input::AbstractString;
            index=:auto,
            save_directives::Bool=false,
            skip_features::Bool=false,
            skip_directives::Bool=true,
            skip_comments::Bool=true)

Create a reader for data in GFF3 format.

The first argument specifies the data source. When it is a filepath that ends with .bgz, it is considered to be block compression file format (BGZF) and the function will try to find a tabix index file (<filename>.tbi) and read it if any. See http://www.htslib.org/doc/tabix.html for bgzip and tabix tools.

Arguments

input: data source (IO object or filepath)

index: path to a tabix file

save_directives: flag to save directive records (which can be accessed with GFF3.directives)

skip_features: flag to skip feature records

skip_directives: flag to skip directive records

skip_comments: flag to skip comment records

GFF3.Record — Method

GFF3.Record(str::AbstractString)

Create a GFF3 record object from str. This function verifies and indexes fields for accessors.

GFF3.Record — Method

GFF3.Record(data::Vector{UInt8})

Create a GFF3 record object from data. This function verifies and indexes fields for accessors. Note that the ownership of data is transferred to a new record object.

GFF3.Record — Method

GFF3.Record()

Create an unfilled GFF3 record.

GFF3.Writer — Type

GFF3.Writer(output::IO)

Create a data writer of the GFF3 file format.

Arguments:

output: data sink

GFF3.attributes — Method

attributes(record::Record, key::String)::Vector{String}

Get the attributes of record with key.

GFF3.attributes — Method

attributes(record::Record)::Vector{Pair{String,Vector{String}}}

Get the attributes of record.

GFF3.content — Method

content(record::Record)::String

Get the content of record. Leading '#' letters are removed.

GFF3.directives — Method

Return all directives that preceded the last GFF entry parsed as an array of strings.

Directives at the end of the file can be accessed by calling close(reader) and then directives(reader).

GFF3.featuretype — Method

featuretype(record::Record)::String

Get the type of record.

GFF3.getfasta — Method

Return a FASTA.Reader initialized to parse trailing FASTA data.

Throws an exception if there is no trailing FASTA, which can be checked using hasfasta.

GFF3.hasfasta — Method

Return true if the GFF3 stream is at its end and there is trailing FASTA data.

GFF3.iscomment — Method

iscomment(record::Record)::Bool

Test if record is a comment record.

GFF3.isdirective — Method

isdirective(record::Record)::Bool

Test if record is a directive record.

GFF3.isfeature — Method

isfeature(record::Record)::Bool

Test if record is a feature record.

GFF3.phase — Method

phase(record::Record)::Int

Get the phase of record.

GFF3.score — Method

score(record::Record)::Float64

Get the score of record.

GFF3.seqend — Method

seqend(record::Record)::Int

Get the end coordinate of record.

GFF3.seqid — Method

seqid(record::Record)::String

Get the sequence ID of record.

GFF3.seqstart — Method

seqstart(record::Record)::Int

Get the start coordinate of record.

GFF3.source — Method

source(record::Record)::String

Get the source of record.

GFF3.strand — Method

strand(record::Record)::GenomicFeatures.Strand

Get the strand of record.